chore: use test package in modules
