+++ modules/kafka/consumer_test.go
@@ -1,4 +1,4 @@
-package kafka
+package kafka_test
 
 import (
 	"testing"
