+++ modules/minio/minio_test.go
@@ -1,4 +1,4 @@
-package minio
+package minio_test
 
 import (
 	"context"
@@ -10,14 +10,15 @@ import (
 	"github.com/minio/minio-go/v7/pkg/credentials"
 
 	"github.com/testcontainers/testcontainers-go"
+	tcminio "github.com/testcontainers/testcontainers-go/modules/minio"
 )
 
 func TestMinio(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx,
+	container, err := tcminio.RunContainer(ctx,
 		testcontainers.WithImage("minio/minio:RELEASE.2024-01-16T16-07-38Z"),
-		WithUsername("thisismyuser"), WithPassword("thisismypassword"))
+		tcminio.WithUsername("thisismyuser"), tcminio.WithPassword("thisismypassword"))
 	if err != nil {
 		t.Fatal(err)
 	}
