+++ modules/mssql/mssql_test.go
@@ -1,4 +1,4 @@
-package mssql
+package mssql_test
 
 import (
 	"context"
@@ -8,14 +8,15 @@ import (
 	_ "github.com/microsoft/go-mssqldb"
 
 	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/modules/mssql"
 	"github.com/testcontainers/testcontainers-go/wait"
 )
 
 func TestMSSQLServer(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx,
-		WithAcceptEULA(),
+	container, err := mssql.RunContainer(ctx,
+		mssql.WithAcceptEULA(),
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -57,7 +58,7 @@ func TestMSSQLServer(t *testing.T) {
 func TestMSSQLServerWithMissingEulaOption(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx)
+	container, err := mssql.RunContainer(ctx)
 	testcontainers.WithWaitStrategy(
 		wait.ForLog("The SQL Server End-User License Agreement (EULA) must be accepted"))
 
@@ -71,8 +72,8 @@ func TestMSSQLServerWithMissingEulaOption(t *testing.T) {
 func TestMSSQLServerWithConnectionStringParameters(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx,
-		WithAcceptEULA(),
+	container, err := mssql.RunContainer(ctx,
+		mssql.WithAcceptEULA(),
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -114,9 +115,9 @@ func TestMSSQLServerWithConnectionStringParameters(t *testing.T) {
 func TestMSSQLServerWithCustomStrongPassword(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx,
-		WithAcceptEULA(),
-		WithPassword("Strong@Passw0rd"),
+	container, err := mssql.RunContainer(ctx,
+		mssql.WithAcceptEULA(),
+		mssql.WithPassword("Strong@Passw0rd"),
 	)
 	if err != nil {
 		t.Fatal(err)
@@ -150,11 +151,11 @@ func TestMSSQLServerWithCustomStrongPassword(t *testing.T) {
 func TestMSSQLServerWithInvalidPassword(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx,
+	container, err := mssql.RunContainer(ctx,
 		testcontainers.WithWaitStrategy(
 			wait.ForLog("Password validation failed")),
-		WithAcceptEULA(),
-		WithPassword("weakPassword"),
+		mssql.WithAcceptEULA(),
+		mssql.WithPassword("weakPassword"),
 	)
 
 	if err == nil {
@@ -174,9 +175,9 @@ func TestMSSQLServerWithInvalidPassword(t *testing.T) {
 func TestMSSQLServerWithAlternativeImage(t *testing.T) {
 	ctx := context.Background()
 
-	container, err := RunContainer(ctx,
+	container, err := mssql.RunContainer(ctx,
 		testcontainers.WithImage("mcr.microsoft.com/mssql/server:2022-RTM-GDR1-ubuntu-20.04"),
-		WithAcceptEULA(),
+		mssql.WithAcceptEULA(),
 	)
 	if err != nil {
 		t.Fatalf("Failed to create the container with alternative image: %s", err)
